services/pollgen-gemini/requirements.txt
streamlit
chromadb
google-generativeai
python-dotenv
sentence-transformers
